+++ MatchingEngine/MatchingEngine.hpp
@@ -14,7 +14,7 @@
 
 class MatchingEngine {
 private:
-    std::unordered_map<uint64_t , OrderBook> orderBook; // sybol assiciated orderbook
+    std::unordered_map<uint8_t , OrderBook> orderBook; // sybol assiciated orderbook
     //std::unordered_map<uint64_t, std::vector<OrderEvent>> orderHistory; // orderid, Orders vector    
     std::unordered_set<uint64_t> orderIds;
     TQueue<Command> qe;
@@ -29,7 +29,7 @@ public:
 
     void PrintTrades();
 
-    void RecordTrade(Order& incoming, Order& recieving, uint64_t quantity, uint64_t price);
+    void RecordTrade(Order& incoming, Order& recieving, uint32_t quantity,uint32_t  price);
 
     template <typename OppositeBook, typename Compare>
     bool CanFullFillOrder(const Order& order, OppositeBook& oppositeBook, Compare comp);
@@ -40,16 +40,16 @@ public:
     void MatchOrder(Order& order, OppositeBook& oppositeBook, Compare comp);
 
     void ProcessOrder(Order& order);
-    void RecordOrderEvent(Order& order, Status newStatus, uint64_t execquantity = 0, uint64_t origquantity = -1);
+    void RecordOrderEvent(Order& order, Status newStatus, uint32_t execquantity = 0, uint32_t origquantity = -1);
     void PrintOrderHistory();
     void ProcessBUY(Order& order);
     void ProcessSELL(Order& order);
-    bool DuplicateOrder(int orderId);
-    void CancelOrder(uint64_t Orderid, uint64_t symbol);//just a function to delegate to orderbook cancel order thing
-    void ModifyOrder(uint64_t orderId, uint64_t symbol, uint64_t newprice = -1, uint64_t newquantity = -1, char newside = '\0'); // -1 mean no value thats why by default
+    bool DuplicateOrder(uint64_t orderId);
+    void CancelOrder(uint64_t Orderid, uint8_t symbol);//just a function to delegate to orderbook cancel order thing
+    void ModifyOrder(uint64_t orderId, uint8_t symbol, uint32_t newprice = -1, uint32_t newquantity = -1, char newside = '\0'); // -1 mean no value thats why by default
 
     //getters
-    OrderBook& GetOrderBook(const uint64_t& symbol) {
+    OrderBook& GetOrderBook(const uint8_t& symbol) {
         return orderBook[symbol];
     }
